+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.0.10] - tbd
+## [1.0.10] - 2024-04-08
 
 ### Added
 - CosMX reader with image stitching (experimental)
@@ -8,6 +8,7 @@
 - Minimum number of transcripts per patch set to 4000 (#41)
 - Config files refactoring (configs added or renamed)
 - Readers refactoring
+- Section with error during report are not displayed (instead of throwing an error)
 
 ## [1.0.9] - 2024-04-03